# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Kartala Task Management API",
        "title": "Kartala Task Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate user and return JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.LoginParameter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Get a new access token using a valid refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "refresh_token": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Create a new user account with username and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_auth.RegisterParameter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (Email or Username already exists)",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerBadRequestResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "email": "must be a valid email address"
                    }
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 400
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request parameters"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerLoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerLoginResponseData"
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 200
                        },
                        "message": {
                            "type": "string",
                            "example": "Login successful"
                        },
                        "status": {
                            "type": "string",
                            "example": "success"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerLoginResponseData": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 900
                },
                "refresh_token": {
                    "type": "string",
                    "example": "def456..."
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerRegisterResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 201
                        },
                        "message": {
                            "type": "string",
                            "example": "Registration successful"
                        },
                        "status": {
                            "type": "string",
                            "example": "success"
                        }
                    }
                }
            }
        },
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 401
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid or expired session"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
...
        "internal_auth.LoginParameter": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "internal_auth.RegisterParameter": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and then your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}